Added modal on to reservation system
